REFACTOR : 인터페이스, package 변경, validate
BaseBallMachine 인터페이스화
package 구조 변경
validate 함수 분리
