+++ docs/README.md
@@ -27,6 +27,9 @@
 - [x] UI Interface 분리
 - [x] Ball, Strike를 구분할 수 있는 새로운 Class 생성
 - [x] 햇갈릴 수 있는 변수명 변경
+- [x] BaseBallMachine 인터페이스화
+- [x] package 구조 변경
+- [x] validate 함수 분리
 
 
 <br>
@@ -83,3 +86,9 @@
 * UI의 경우 지금은 Console 입력을 받지만 나중에는 사용자 입력을 받는 경우도 있을 것이라고 보고 UI를 Interface로 분리하고, 생성하는 시점에 정할 수 있도록 만들었습니다.
 * 함수를 사용해서, 그냥 보기 어려웠던 BallCount를 세는 부분을 가독성 있게 변경
 * 함수명, 변수명 중 햇갈릴 수 있는 함수, 변수명에 대해서 가독성이 있게 변경함
+
+**2023-10-24**
+
+* BaseballMachine 도 난이도에 따라 변경될 수 있지 않을까 하는 생각에 인터페이스화 함.
+* package 구조를 변경해서 가장 상단에 Application 하나를 두고 그 이하는 machine패키지를 사용해서 분리함.
+* validate의 경우 함수를 따로 분리하고, 각자의 메서드를 책임에 맞게 분리함.
