+++ src/main/java/baseball/Application.java
@@ -1,13 +1,14 @@
 package baseball;
 
-import baseball.ui.ConsoleUI;
-import baseball.ui.UI;
+import baseball.machine.BaseballMachine;
+import baseball.machine.NormalBaseballMachine;
+import baseball.machine.ui.ConsoleUI;
+import baseball.machine.ui.UI;
 
 public class Application {
     public static void main(String[] args) {
-        // TODO: 프로그램 구현
         UI ui = new ConsoleUI();
-        BaseballMachine machine = new BaseballMachine(ui);
+        BaseballMachine machine = new NormalBaseballMachine(ui);
 
         machine.start();
