+++ src/main/java/baseball/machine/BaseballMachine.java
@@ -1,0 +1,7 @@
+package baseball.machine;
+
+public interface BaseballMachine {
+
+    void start();
+
+}
